requirements.txt
click>=8.1.0
requests>=2.31.0
rich>=13.7.0
pandas>=2.1.0
pyyaml>=6.0
jinja2>=3.1.2
plotly>=5.18.0
python-dateutil>=2.8.2
openpyxl>=3.1.0
tabulate>=0.9.0
